requirements.txt
prettytable==2.0.0
numpy==1.26.4
torch==2.2.2
gdown==3.12.2
imageio==2.9.0
sconf==0.2.3
torchvision==0.17.2
tqdm==4.51.0
tensorboard==2.3.0
Pillow==8.1.0
tensorboardX==2.1
pydantic==2.9.2
